Fix prompt
