+++ src/deal_command.c
@@ -1,0 +1,35 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   deal_command.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cpak <cpak@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/04/06 14:39:03 by cpak              #+#    #+#             */
+/*   Updated: 2022/04/06 14:39:04 by cpak             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int deal_command(char *str, char *envp[])
+{
+    pid_t   pid;
+    int     status;
+    int     fds[2];
+
+    pipe(fds);
+    pid = fork();
+    if (pid == 0)
+    {
+        g_mini.cmd_len = 0;
+        if (parse_command(str) == 0)
+            exit(0);
+        // read_all_cmd();  // 입력된 명령어 리스트 출력
+        pipex(g_mini.cmd_len - 1, envp);
+        exit(0);
+    }
+    else
+        waitpid(pid, &status, 0);
+    return (1);
+}
